Unit test + library
fdox now has an internal c library, sort of. Just a set of sources and includes.
Started working on unit test - currently only testing for `find_applet`
main.c is now part of the project, and no longer auto-generated.

applets.c is autogenerated, but I have a bug in the `configure.sh` script
which treats files of the library as commands, so it needs to be edited
manually after configuration, otherwise the app does not compile.

to test:
`make test`