scripts to recreate an epic dos drive for development
lets bring devops to dos?

This script downloads a set of packages for development,
a few compilers, 4DOS, a file manager (and more to come).

The idea is that you setup the created disk as C: in your
dosbox, and D: as your source - so you can develop as easy as
possible.