Copy : open file in binary mode
If we don't open the file in binary mode, then newlines are modified on
windows (the code reads \r\n and emits just \r.

I am not even trying to enable a binary mode like the MSDOS command, as
this is stupid. If you do line convertions - this is not the tool for
you. (I might do a separate tool later on).