Added initial support for time
The shell now prints time. It needs to be coded better (`atoi()` is
bad), and I need to support setting the time... but it display it (no
AM/PM support yet).

This was done easily by merging the code into date (which should be
called `datetime.c` soon) - most of the code is shared anyway.c

TODO:
- port away `atoi`
- support for printing AM/PM
- support for setting time
- support for setting AM/PM time