Remove example command
This was not working, well more or less. It triggered the args
parsing API, so it was time well spent. Bye!

If you want an example of how to build a new command, having a
template, you should from now on just copy `dos/type.c` as
a small example (next commit...).