implementation of "set" use proper $PROMPT
This prooved to be a challenge: in theory I should use the command
`setenv()`. However, it is not available under TC2.02. Instead I should
use somehow `putenv()`.

The problem with `putenv()` is that it uses the raw buffer for the value.
So, I cannot use a variable on stack, nor a static one. I need to
maintain a list of "known/modified" variables, and use that as the
pointer to the string sent to `putenv()`. This is not trivial as I also
need to free older values.. etc.

I am using a list of 20 user values. How about using a linked list some
day?

The main command uses now the correct environment variable (%PROMPT%) to
display the user prompt.

Also - commamand how has a simple function to execute a single line
(which will be used to script the internal unit-tests soon).  I used it
to feed a list of `set` commands and test how the internal API works.