Workaround a bug in DJGPP
I found a bug in DJGPP (gcc9 and gcc10).  I will need to reproduce in a
small scale, but the essential bug is:

```
char c = 'a';
c = tolower(c); // now c=0 ...
```

The workarround is :

```
char c1 = 'a';
char c2 = tolower(c1);
```

I need to overcome this in the library itself,but for now it's done in
all the applets. Code works, we continue.