Lets start working on PacificC support
This compiler is mostly compatible with TurboC. But... its runtime is
missing some functions. Also - some includes have different names.

Code does not fully compile yet, working on it. I ported some of the
fixes from this branch to the main branch. This branch will follow main
for the following time, until I feel that its ready to get merged in.

Functions missing:
* glog will need another implementation
* put_env is not existant
* snfprintf... well, its expected