Lets move all keyboard handling to functions
This is becomming a real life library. You need to initialize the
library (on Posix, in Windows/DOS the init function is nil). Then you
need to initialize a session and then run it.

All cursor movement, or text modifications - is done via dedicated
funcions. Which means in the future, I can add an undo/redo stack.

I am still providing a simpler API - which internally uses sessions.