Control WIP
Theory - we should get to catch "singal 5" and then we should
be ready to read control+c.

In DOS this works, I am still unable to test this on Linux, as control-c
send us the kill signal. I hope to fix this by pressing control+c twice to
`abort()`.

Leaving this as a test branch for now.